+++ ui/src/components/form-view/QasFormView.vue
@@ -46,7 +46,7 @@ import { viewMixin } from '../../mixins'
 import debug from 'debug'
 import { extend } from 'quasar'
 import { getAction } from '@bildvitta/store-adapter'
-import { isEqualWith } from 'lodash-es'
+import isEqualWith from 'lodash-es/isEqual'
 import { onBeforeRouteLeave } from 'vue-router'
 
 const log = debug('asteroid-ui:qas-form-view')